Introduce GFC_STD_UNSIGNED.
This patch creates an unsigned "standard" for the
gfc_option.allow_std field.

One of the main reason why people want UNSIGNED for Fortran is
interfacing for C.

This is a preparation for further work on the ISO_C_BINDING constants.
That, we do via iso-c-binding.def , whose last field is a standard
for the constant to be defined for the standard in question, which is
then checked.  I could try and invent a different method for this,
but I'd rather not.

gcc/fortran/ChangeLog:

	* intrinsic.cc (add_functions): Convert uint and
	selected_unsigned_kind to GFC_STD_UNSIGNED.
	(gfc_check_intrinsic_standard): Handle GFC_STD_UNSIGNED.
	* libgfortran.h (GFC_STD_UNSIGNED): Add.
	* options.cc (gfc_post_options): Set GFC_STD_UNSIGNED
	if -funsigned is set.