passes: Allow for second param for NEXT_PASS
Right now we currently only support 1 parameter for each pass in NEXT_PASS.
We also don't error out if someone tries to use more than 1.
This adds support for more than one but only to a max of max_number_args
(which is currently 2).
In the next patch, this will be used for DCE, adding a new parameter.

Bootstrapped and tested on x86_64-linux-gnu.

gcc/ChangeLog:

	* gen-pass-instances.awk (END): Handle processing
	of multiple arguments to NEXT_PASS. Also error out
	if using more than max_number_args (2).
	* pass_manager.h (NEXT_PASS_WITH_ARG2): New define.
	* passes.cc (NEXT_PASS_WITH_ARG2): New define.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>