dce: Use a base common base class for pass_cd_dce and pass_dce
The classes pass_dce and pass_cd_dce share the same mechansim for their
params and almost the same execute functionality so let's create a new
base class which will be used for these two classes and move the common
code into the same one.

Note update_address_taken_p was updated to be a NSDMI instead of initializing
it explicitly in the constructor.

Bootstrapped and tested on x86_64-linux-gnu.

gcc/ChangeLog:

	* tree-ssa-dce.cc (tree_ssa_dce): Remove.
	(tree_ssa_cd_dce): Remove.
	(class pass_dce_base): New class.
	(class pass_dce): Use pass_dce_base as the base class.
	(class pass_cd_dce): Likewise.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>