Only do switch bit test clustering when multiple labels point to same bb
The bit cluster code generation strategy is only beneficial when
multiple case labels point to the same code. Do a quick check if
that is the case before trying to cluster.

This fixes the switch part of PR117091 where all case labels are unique
however it doesn't address the performance problems for non unique
cases.

gcc/ChangeLog:

	PR middle-end/117091
	* gimple-if-to-switch.cc (if_chain::is_beneficial): Update
	find_bit_test call.
	* tree-switch-conversion.cc (bit_test_cluster::find_bit_tests):
	Get max_c argument and bail out early if all case labels are
	unique.
	(switch_decision_tree::compute_cases_per_edge): Record number of
	targets per label and return.
	(switch_decision_tree::analyze_switch_statement): ... pass to
	find_bit_tests.
	* tree-switch-conversion.h: Update prototypes.