doc: remove outdated C++ Concepts section
This was added as part of the initial Concepts TS implementation and
reflects an early version of the Concepts TS paper, which is very
different from standard C++20 concepts (and even from more recent
versions of the Concepts TS, support for which we deprecated in GCC 14
and removed for GCC 15).  So there's not much to salvage from this
section besides the __is_same trait documentation which we can
conveniently move to the previous Type Traits section.

gcc/ChangeLog:

	* doc/extend.texi (C++ Concepts): Remove section.  Move
	__is_same documentation to the previous Type Traits section.

Reviewed-by: Jason Merrill <jason@redhat.com>