simplify-rtx: Simplify ROTATE:HI (X:HI, 8) into BSWAP:HI (X)
With recent patch to improve detection of vector rotates at RTL level
combine now tries matching a V8HImode rotate by 8 in the example in the
testcase.  We can teach AArch64 to emit a REV16 instruction for such a rotate
but really this operation corresponds to the RTL code BSWAP, for which we
already have the right patterns.  BSWAP is arguably a simpler representation
than ROTATE here because it has only one operand, so let's teach simplify-rtx
to generate it.

With this patch the testcase now generates the simplest form:
.L2:
        ldr     q31, [x1, x0]
        rev16   v31.16b, v31.16b
        str     q31, [x0, x2]
        add     x0, x0, 16
        cmp     x0, 2048
        bne     .L2

instead of the previous:
.L2:
        ldr     q31, [x1, x0]
        shl     v30.8h, v31.8h, 8
        usra    v30.8h, v31.8h, 8
        str     q30, [x0, x2]
        add     x0, x0, 16
        cmp     x0, 2048
        bne     .L2

IMO ideally the bswap detection would have been done during vectorisation
time and used the expanders for that, but teaching simplify-rtx to do this
transformation is fairly straightforward and, unlike at tree level, we have
the native RTL BSWAP code.  This change is not enough to generate the
equivalent sequence in SVE, but that is something that should be tackled
separately.

Bootstrapped and tested on aarch64-none-linux-gnu.

Signed-off-by: Kyrylo Tkachov <ktkachov@nvidia.com>

gcc/

	* simplify-rtx.cc (simplify_context::simplify_binary_operation_1):
	Simplify (rotate:HI x:HI, 8) -> (bswap:HI x:HI).

gcc/testsuite/

	* gcc.target/aarch64/rot_to_bswap.c: New test.